+++ src/components/templates/AboutPageTemplate.tsx
@@ -1,0 +1,26 @@
+import React from 'react'
+import Content, { HTMLContent } from '../Content'
+
+const AboutPageTemplate:React.FC<any> = (props) => {
+const { title, content, contentComponent } = props
+  const PageContent = contentComponent || Content
+
+  return (
+    <section className="section section--gradient">
+      <div className="container">
+        <div className="columns">
+          <div className="column is-10 is-offset-1">
+            <div className="section">
+              <h2 className="title is-size-3 has-text-weight-bold is-bold-light">
+                {title}
+              </h2>
+              <PageContent className="content" content={content} />
+            </div>
+          </div>
+        </div>
+      </div>
+    </section>
+  )
+}
+
+export default AboutPageTemplate
