+++ src/components/templates/index.tsx
@@ -1,0 +1,4 @@
+export { default as IndexPageTemplate } from './IndexPageTemplate'
+export { default as AboutPageTemplate } from './AboutPageTemplate'
+export { default as BlogPostTemplate } from './BlogPostTemplate'
+export { default as ProductPageTemplate } from './ProductPageTemplate'
