+++ src/container/about-page.tsx
@@ -1,0 +1,32 @@
+import React from 'react'
+import { graphql } from 'gatsby'
+import Layout from '../components/Layout'
+import { HTMLContent } from '../components/Content'
+import { AboutPageTemplate } from '../components/templates';
+
+const AboutPage = ({ data }) => {
+  const { markdownRemark: post } = data
+
+  return (
+    <Layout>
+      <AboutPageTemplate
+        contentComponent={HTMLContent}
+        title={post.frontmatter.title}
+        content={post.html}
+      />
+    </Layout>
+  )
+}
+
+export default AboutPage
+
+export const aboutPageQuery = graphql`
+  query AboutPage($id: String!) {
+    markdownRemark(id: { eq: $id }) {
+      html
+      frontmatter {
+        title
+      }
+    }
+  }
+`
